% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_scopes.R
\name{compare_scopes}
\alias{compare_scopes}
\title{Print metrics in three scope comparable columns}
\usage{
compare_scopes(data)
}
\arguments{
\item{data}{A data.frame with metrics data.}
}
\description{
Print metrics in three scope comparable columns
}
