% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_full.R
\name{filter_full}
\alias{filter_full}
\title{Filter by the full-sample scope for metrics from model run}
\usage{
filter_full(data)
}
\arguments{
\item{data}{Data from metrics().}
}
\description{
Filter by the full-sample scope for metrics from model run
}
