% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_is.R
\name{filter_is}
\alias{filter_is}
\title{Filter by the in-sample scope for metrics from model run}
\usage{
filter_is(data)
}
\arguments{
\item{data}{Data from metrics().}
}
\description{
Filter by the in-sample scope for metrics from model run
}
