% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metrics.R
\name{filter_metrics}
\alias{filter_metrics}
\title{Filter for metrics from model run}
\usage{
filter_metrics(data, ...)
}
\arguments{
\item{data}{Data from metrics().}

\item{...}{Expression for the filter.}
}
\description{
Filter for metrics from model run
}
