% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_oos.R
\name{filter_oos}
\alias{filter_oos}
\title{Filter by the out-of-sample scope for metrics from model run}
\usage{
filter_oos(data)
}
\arguments{
\item{data}{AData from metrics().}
}
\description{
Filter by the out-of-sample scope for metrics from model run
}
