% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_step.R
\name{filter_step}
\alias{filter_step}
\title{Filter by the specified step}
\usage{
filter_step(data, step)
}
\arguments{
\item{data}{A data.frame with step feature in it.}

\item{step}{An integer with the step to filter.}
}
\description{
Filter by the specified step
}
