% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binance_ohlcv_data.R
\name{get_binance_ohlcv_data}
\alias{get_binance_ohlcv_data}
\title{Get OHLCV data from Binance}
\usage{
get_binance_ohlcv_data(symbol, timeframe)
}
\arguments{
\item{symbol}{A string with the symbol ("BTC", "ETH", ...).}

\item{timeframe}{A string with the timeframe ("15m", "1h", "4h", ...).}
}
\value{
This function retrieves OHLCV data from Binance and saves in ./datasets.
}
\description{
Get OHLCV data from Binance
}
