% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_steps.R
\name{get_model_steps}
\alias{get_model_steps}
\title{Get all the step combinations from model}
\usage{
get_model_steps(model)
}
\arguments{
\item{model}{An R object with model or a string with the name of the YAML file describing the model.}
}
\value{
The function returns a list with 'n_steps' (the number of steps) and 'steps' (a data.frame with the combinations).
}
\description{
The function get_model_steps() reads the steps section in the model or the model YAML file and builds a grid with all the possible combinations.
}
