% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_ohlcv_data.R
\name{get_yahoo_ohlcv_data}
\alias{get_yahoo_ohlcv_data}
\title{Get OHLCV data from Yahoo}
\usage{
get_yahoo_ohlcv_data(symbol, from = "2000-01-01")
}
\arguments{
\item{symbol}{A string with the symbol ("SPY", "QQQ", "NVDA", "TSLA", "EUR=X", "GBP=X", ...).}

\item{from}{A string with the start date in YYYY-mm-dd format.}
}
\value{
This function retrieves OHLCV data from Yahoo and saves in ./datasets.
}
\description{
Get OHLCV data from Yahoo
}
