% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_oos_ratios.R
\name{is_oos_ratios}
\alias{is_oos_ratios}
\title{Print in-sample to out-os-sample ratio}
\usage{
is_oos_ratios(data, assess = TRUE)
}
\arguments{
\item{data}{A data.frame with metrics data.}

\item{assess}{A boolean for ratio assessment.}
}
\description{
Print in-sample to out-os-sample ratio
}
