% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot balance}
\usage{
plot_balance(step, show_source = FALSE)
}
\arguments{
\item{step}{An integer with the model step to plot.}

\item{show_source}{A boolean indicating whether or not to plot the data source. If TRUE, the price data from the model will be plotted in a secondary axis.}
}
\description{
Plot balance
}
