% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance_excursions.R
\name{plot_balance_excursions}
\alias{plot_balance_excursions}
\title{Plot balance and excursions}
\usage{
plot_balance_excursions(step = 1)
}
\arguments{
\item{step}{An integer with the model step to plot.}
}
\description{
Plot balance and excursions
}
