% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark.R
\name{plot_benchmark}
\alias{plot_benchmark}
\title{Plot a comparison of the specified step in model with the Buy & Hold strategy}
\usage{
plot_benchmark(step)
}
\arguments{
\item{step}{n integer with the balance step to plot.}
}
\description{
Plot a comparison of the specified step in model with the Buy & Hold strategy
}
