% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_best_metrics_i.R
\name{plot_best_metrics_i}
\alias{plot_best_metrics_i}
\title{Plot balances for best metrics interactively}
\usage{
plot_best_metrics_i()
}
\description{
Plot balances for best metrics interactively
}
