% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_excursions.R
\name{plot_excursions}
\alias{plot_excursions}
\title{Plot excursions}
\usage{
plot_excursions(step = 1)
}
\arguments{
\item{step}{An integer with the model step to plot.}
}
\description{
Plot excursions
}
