% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_max_excursions.R
\name{plot_max_excursions}
\alias{plot_max_excursions}
\title{Plot maximum excursions}
\usage{
plot_max_excursions(step = 1, show_sum_return = FALSE)
}
\arguments{
\item{step}{An integer with the model step to plot.}

\item{show_sum_return}{A logical value. If TRUE, sum of returns line is shown.}
}
\description{
Plot maximum excursions
}
