% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monte_carlo.R
\name{plot_monte_carlo}
\alias{plot_monte_carlo}
\title{Plot Monte Carlo simulations}
\usage{
plot_monte_carlo(
  step = 1,
  scope = "is",
  samples = 500,
  replace = TRUE,
  oos_zoom = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{step}{An integer with a valid step in balances.}

\item{scope}{A string for the scope taking "is" (in-sample) or "oos" (out-of-sample). Default "is".}

\item{samples}{An integer with the number of iterations. Default 500.}

\item{replace}{Logical value for resampling with or without replacement. Default TRUE.}

\item{oos_zoom}{Logical value for zooming out-of-sample region. Default TRUE.}

\item{verbose}{Logical value for showing the progress. Default TRUE.}
}
\description{
Plot Monte Carlo simulations
}
