% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trades.R
\name{plot_trades}
\alias{plot_trades}
\title{Plot OHLC bars and trades}
\usage{
plot_trades(model_file, step = 1, from = NULL, to = NULL, show_trades = TRUE)
}
\arguments{
\item{model_file}{A string with the name of the YAML file describing the model.}

\item{step}{An integer with the model step to plot.}

\item{from}{A string with date formatted as yyyy-mm-dd.}

\item{to}{A string with date formatted as yyyy-mm-dd.}

\item{show_trades}{A boolean indicating whether or not show the trades.}
}
\description{
Plot OHLC bars and trades
}
