% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_returns.R
\name{print_returns}
\alias{print_returns}
\title{Print returns}
\usage{
print_returns(step, pretty = TRUE)
}
\arguments{
\item{step}{A integer with the model step to print.}

\item{pretty}{Logical value for showing pretty table. Default TRUE.}
}
\description{
Print returns
}
