% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs a model on OHLCV data}
\usage{
run_model(model_file, steps = NULL, preview = FALSE, verbose = TRUE)
}
\arguments{
\item{model_file}{A string with the name of the YAML file describing the model.}

\item{steps}{An integer vector with the subset of steps to be run. If NULL, all the steps are run. Default NULL.}

\item{preview}{A logical value. If TRUE, a performance preview for each step is plotted. Default TRUE.}

\item{verbose}{A logical value. If TRUE, informative messages are printed. Default TRUE.}
}
\value{
Balance, metrics and other data is stored in a temporary database for further analysis.
}
\description{
A model is a YAML file containing information about entry/exit strategy rules and indicators setup among other data. The run_model() function performs a complete backtest by taking a model and executing it on the provided OHLCV data.
}
