% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signals}
\alias{signals}
\title{Signals}
\usage{
signals(data, step)
}
\arguments{
\item{data}{A data.frame with balance data.}

\item{step}{A integer with the model step to get timings.}
}
\description{
Signals
}
