% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_metrics.R
\name{sort_metrics}
\alias{sort_metrics}
\title{Sort metrics}
\usage{
sort_metrics(data, ...)
}
\arguments{
\item{data}{Metrics data from metrics().}

\item{...}{Variable to sort.}
}
\description{
Sort metrics
}
